PORTED_MODEL_CODE/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "industrial-smoke-plume-detection"
version = "0.2.0"
description = "Industrial smoke plume detection and segmentation from Sentinel-2 imagery"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "matplotlib>=3.7",
    "rasterio>=1.3",
    "scikit-learn>=1.3",
    "shapely>=2.0",
    "tqdm>=4.65",
    "tensorboard>=2.14",
    "torch>=2.2",
    "torchvision>=0.17",
    "lightning>=2.2",
    "torchmetrics>=1.3",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "pip>=26.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-timeout>=2.3",
    "pre-commit>=3.6",
    "code-review-graph>=2.3",
]
notebooks = [
    "jupyter>=1.0",
    "ipykernel>=6.29",
]

[tool.setuptools.packages.find]
where = ["src"]

# --- PyTorch CUDA wheels ---------------------------------------------------
# torch + torchvision resolve from the PyTorch cu124 index, not default PyPI.
# Change the index URL + source entries if you need a different CUDA version
# (e.g. cu121, cu126, cu128). Wheels bundle their own CUDA runtime, so this
# works without a system CUDA toolkit installed — driver is only needed at
# runtime to actually use the GPU.
[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]
torchvision = [{ index = "pytorch-cu124" }]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
ignore = ["E501"]  # line-length handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers --cov=src/smoke_detection --cov-branch --cov-report=term-missing -m 'not slow and not gpu and not e2e' --timeout=30"
markers = [
    "slow: tests > ~2s (pretrained weight load, overfitting sanity loops)",
    "gpu: requires CUDA; auto-skipped when unavailable",
    "e2e: end-to-end pipelines (fast_dev_run, CLI, ckpt round-trip)",
]
# Silence environmental warnings that are expected in the test environment:
#   - CPU accelerator is pinned deliberately (even when CUDA is available).
#   - num_workers=0 is deliberate for fast test setup.
#   - Tiny synthetic val batches can be all-negative, producing a torchmetrics
#     "No positive samples in targets" warning; real runs don't hit this.
filterwarnings = [
    "ignore:GPU available but not used:UserWarning",
    "ignore:.*does not have many workers.*:UserWarning",
    "ignore:No positive samples in targets.*:UserWarning",
]

[tool.coverage.run]
branch = true
source = ["src/smoke_detection"]
omit = ["src/smoke_detection/__init__.py", "*/__main__.py"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "raise NotImplementedError", "if TYPE_CHECKING:"]
